+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add encoding as CLI param.
 - Add `--offline` option, and use it internally.
 - Fix the CLI parameter parsing: Either `--local_directory` or `--tsv_paths` must be provided.
+- Use the SciCrunch resolver for RR IDs.
 - More helpful message if decoding error.
 
 ## v0.0.7 - 2021-01-13